Cluster uint16 to uint32

+++ contracts/Cluster/Cluster.sol
@@ -12,13 +12,13 @@ contract Cluster is Ownable, ICluster {
     // ************ //
 
     /// @notice returns the current LayerZero chain id
-    uint16 public lzChainId;
+    uint32 public lzChainId;
     /// @notice returns true if an address is marked as an Editor
     /// @dev editors can update contracts' whitelist status
     mapping(address editor => bool status) public isEditor;
     /// @notice returns the whitelist status for an address
     /// @dev LZ chain id => contract => status
-    mapping(uint16 lzChainId => mapping(address _contract => bool status))
+    mapping(uint32 lzChainId => mapping(address _contract => bool status))
         private _whitelisted;
 
     /// @notice event emitted when LZ chain id is updated
@@ -32,7 +32,7 @@ contract Cluster is Ownable, ICluster {
     /// @notice event emitted when a contract status is updated
     event ContractUpdated(
         address indexed _contract,
-        uint16 indexed _lzChainId,
+        uint32 indexed _lzChainId,
         bool indexed _oldStatus,
         bool _newStatus
     );
@@ -42,7 +42,7 @@ contract Cluster is Ownable, ICluster {
     // ************** //
     error NotAuthorized();
 
-    constructor(uint16 _lzChainId, address _owner) {
+    constructor(uint32 _lzChainId, address _owner) {
         lzChainId = _lzChainId;
         transferOwnership(_owner);
     }
@@ -54,7 +54,7 @@ contract Cluster is Ownable, ICluster {
     /// @param _lzChainId LayerZero chain id
     /// @param _addr the contract's address
     function isWhitelisted(
-        uint16 _lzChainId,
+        uint32 _lzChainId,
         address _addr
     ) external view override returns (bool) {
         if (_lzChainId == 0) {
@@ -73,7 +73,7 @@ contract Cluster is Ownable, ICluster {
     /// @param _addresses the contracts addresses
     /// @param _status the new whitelist status
     function batchUpdateContracts(
-        uint16 _lzChainId,
+        uint32 _lzChainId,
         address[] memory _addresses,
         bool _status
     ) external override {
@@ -102,7 +102,7 @@ contract Cluster is Ownable, ICluster {
     /// @param _addr the contract's address
     /// @param _status the new whitelist status
     function updateContract(
-        uint16 _lzChainId,
+        uint32 _lzChainId,
         address _addr,
         bool _status
     ) external override {
@@ -128,7 +128,7 @@ contract Cluster is Ownable, ICluster {
     // ********************* //
     /// @notice updates LayerZero chain id
     /// @param _lzChainId the new LayerZero chain id
-    function updateLzChain(uint16 _lzChainId) external onlyOwner {
+    function updateLzChain(uint32 _lzChainId) external onlyOwner {
         emit LzChainUpdate(lzChainId, _lzChainId);
         lzChainId = _lzChainId;
     }
